refactor: LottoCalculation 클래스 삭제
